Ma'lumot o'chirishda agar boshqa tablitsalarga bog'langan columnlar b… …o'lsa o'chirish mumkin emasligi haqidagi toastr chiqadi udalit qilishda

+++ app/Http/Controllers/CategoryController.php
@@ -80,6 +80,14 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $model = Category::where('step', 0)->find($id);
+        if(!empty($model->subcategory)){
+            if(!$model->subcategory->isEmpty()){
+                return redirect()->back()->with('error', translate('You cannot delete this category because it has subcategories'));
+            }
+        }
+        if($model->product){
+            return redirect()->back()->with('error', translate('You cannot delete this category because it has products'));
+        }
         $model->delete();
         return redirect()->route('category.index')->with('status', translate('Successfully deleted'));
     }
